move litellm_utils => litellm_lib
